separated each page into its own file to render each file based on nav button clicked
